pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screen-agent"
version = "0.2.0"
description = "Give AI agents eyes and hands on the desktop. An MCP server for screen perception and interaction with multi-backend input chain."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Chris Wu", email = "yichenwujob@gmail.com" }]
keywords = ["mcp", "screen", "ai", "agent", "claude", "computer-use", "automation", "accessibility"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "typer>=0.12.0",
    "pynput>=1.7.0",
]

[project.optional-dependencies]
# macOS native backends (recommended)
macos = [
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Vision>=10.0",
]
# HTTP transport
api = ["fastapi>=0.110.0", "uvicorn>=0.29.0"]
# Development
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "ruff>=0.4.0"]

[project.urls]
Homepage = "https://github.com/chriswu727/screen-agent"
Repository = "https://github.com/chriswu727/screen-agent"
Issues = "https://github.com/chriswu727/screen-agent/issues"

[project.scripts]
screen-agent = "screen_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/screen_agent"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
